+++ Source/Encoding/x265Enc.vb
@@ -1048,12 +1048,10 @@ Public Class x265Params
                     Scenecut,
                     New NumParam() With {.Switch = "--scenecut-bias", .Text = "Scenecut Bias", .Init = 5, .Config = {0, 100, 1, 1}},
                     New NumParam() With {.Switch = "--radl", .Text = "Radl"},
-                    New NumParam() With {.Switch = "--hist-threshold", .Text = "Hist Threshold", .Init = 0.03, .Config = {0, 1, 0.01, 2}},
                     Ref)
-                Add("Slice Decision 2", MinKeyint, Keyint, Bpyramid, OpenGop, IntraRefresh,
+                Add("Slice Decision 2", MinKeyint, Keyint, OpenGop, Bpyramid, IntraRefresh,
                     New BoolParam() With {.Switch = "--fades", .Text = "Detection and handling of fade-in regions"},
-                    New BoolParam() With {.Switch = "--hist-scenecut", .NoSwitch = "--no-hist-scenecut", .Text = "Scenecut detection using luma edge and chroma histograms"},
-                    New BoolParam() With {.Switch = "--traditional-scenecut", .NoSwitch = "--no-traditional-scenecut", .Init = True, .Text = "Traditional scenecut detection using intra and inter cost when '--hist-scenecut` is used."})
+                    New BoolParam() With {.Switch = "--hist-scenecut", .NoSwitch = "--no-hist-scenecut", .Text = "Scenecut detection using luma edge and chroma histograms"})
                 Add("Performance",
                     New StringParam With {.Switch = "--pools", .Switches = {"--numa-pools"}, .Text = "Pools"},
                     New NumParam With {.Switch = "--slices", .Text = "Slices", .Init = 1},